refactor(panel): one way to read the view, one home for the coercion
The second code round, three findings:

- `PanelProvider` had both `this.held` and a `view` getter aliasing it. Two
  ways to read one thing is two ways for a later method to bypass the handle's
  invariants; every read goes through `this.held.view` now.
- `asText` lived in `escapeHtml.ts` and was imported into `roundsLog.ts` to
  sanitise a file PATH — an HTML module pulled into path parsing. It is its
  own module, imported by all three.
- The webview-owner scan read only the top level of `src/`, so an owner added
  under `src/foo/` would never be discovered and the suite would stay green
  while it held a disposed view. It recurses.

Suite: 779 passed, 0 failed.
